Utvider valideringstest med flere tester
